requirements.txt
fastapi
uvicorn
ultralytics
pillow
numpy
pyngrok
